refactor: move routing code from main.rs
- move `Mount` to router.rs
- make middlewares mod
    - move Logger middleware to middlewares
